Isolate Qwen3 native accelerators
